requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==4.1
django-rest-framework==0.1.0
django-userforeignkey==0.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
